% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/covid.R
\docType{data}
\name{covid}
\alias{covid}
\title{global covid-19 epidemic}
\format{
A data frame with columns:
\describe{
\item{Country/Region}{Country or Region; ex: Mainland China, Romania, Italy}
\item{Province/State}{Province or State; ex: Hubei}
\item{Latitude}{Latitude of the Country}
\item{Longitude}{Longitude of the Country}
\item{Confirmed}{Confirmed number of cases}
\item{Recovered}{Number of recovered patients}
\item{Deaths}{Number of deaths}
\item{Date}{Date of the report}
}
}
\source{
\href{https://www.kaggle.com/andradaolteanu/covid-19-sentiment-analysis-social-networks/?select=covid-19-all.csv}{covid}
}
\usage{
covid
}
\description{
This dataset contains global covid-19 epidemic situations of 215 countries.
}
\keyword{datasets}
