% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/COVIDsen-package.R
\docType{package}
\name{COVIDsen-package}
\alias{COVIDsen}
\alias{COVIDsen-package}
\title{COVIDsen: A Shiny application for visualizing the global Covid-19 epidemic}
\description{
The package deployed a shiny application which shows the global COVID-19 epidemic situation.
}
\author{
\strong{Maintainer}: Sen Wang \email{swan0165@student.monash.edu} (\href{https://orcid.org/YOUR-ORCID-ID}{ORCID})

}
\keyword{internal}
