% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/serverFunction.R
\name{filtered}
\alias{filtered}
\title{Create different countries and dates dataset}
\usage{
filtered(data, country, date)
}
\arguments{
\item{data}{selected data}

\item{country}{filtered country}

\item{date}{filtered date}
}
\value{
different countries dataset
}
\description{
Function \code{filtered()} can filter different countries
}
\examples{
\dontrun{
selected(covid, "Australia", "2020-10-01")
}

}
