% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/uiFunction.R
\name{selected}
\alias{selected}
\title{Creates reactive input for Shiny app}
\usage{
selected(id, label)
}
\arguments{
\item{id}{The identity entered by users}

\item{label}{What the user sees when the request is entered}
}
\value{
Creates a reactive slider/picker input for Shiny App
}
\description{
Creates reactive input for Shiny app
}
\examples{
\dontrun{
selected("Country", "Select a Country")
}

}
